added searching and sorting and amenity xUnit testing
