+++ AsyncInn/AsyncInn/Controller/RoomAmenitiesController.cs
@@ -1,0 +1,169 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore;
+using AsyncInn.Data;
+using AsyncInn.Models;
+
+namespace AsyncInn.Constrollers
+{
+    public class RoomAmenitiesController : Controller
+    {
+        private readonly AsyncInnDbContext _context;
+
+        public RoomAmenitiesController(AsyncInnDbContext context)
+        {
+            _context = context;
+        }
+
+        // GET: RoomAmenities
+        public async Task<IActionResult> Index()
+        {
+            var asyncInnDbContext = _context.RoomAmenity.Include(r => r.Amenity).Include(r => r.Room);
+            return View(await asyncInnDbContext.ToListAsync());
+        }
+
+        // GET: RoomAmenities/Details/5
+        public async Task<IActionResult> Details(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var roomAmenities = await _context.RoomAmenity
+                .Include(r => r.Amenity)
+                .Include(r => r.Room)
+                .FirstOrDefaultAsync(m => m.AmenitiesID == id);
+            if (roomAmenities == null)
+            {
+                return NotFound();
+            }
+
+            return View(roomAmenities);
+        }
+
+        // GET: RoomAmenities/Create
+        public IActionResult Create()
+        {
+            ViewData["AmenitiesID"] = new SelectList(_context.Amenity, "ID", "Name");
+            ViewData["RoomID"] = new SelectList(_context.Rooms, "ID", "Name");
+            return View();
+        }
+
+        // POST: RoomAmenities/Create
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Create([Bind("AmenitiesID,RoomID")] RoomAmenities roomAmenities)
+        {
+            if (ModelState.IsValid)
+            {
+                _context.Add(roomAmenities);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            ViewData["AmenitiesID"] = new SelectList(_context.Amenity, "ID", "Name", roomAmenities.AmenitiesID);
+            ViewData["RoomID"] = new SelectList(_context.Rooms, "ID", "Name", roomAmenities.RoomID);
+            return View(roomAmenities);
+        }
+
+
+        //  This is to take out the Update/Edit ability......maybe
+        // GET: RoomAmenities/Edit/5
+        public async Task<IActionResult> Edit(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var roomAmenities = await _context.RoomAmenity.FindAsync(id);
+            if (roomAmenities == null)
+            {
+                return NotFound();
+            }
+            ViewData["AmenitiesID"] = new SelectList(_context.Amenity, "ID", "Name", roomAmenities.AmenitiesID);
+            ViewData["RoomID"] = new SelectList(_context.Rooms, "ID", "Name", roomAmenities.RoomID);
+            return View(roomAmenities);
+        }
+
+        // POST: RoomAmenities/Edit/5
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Edit(int id, [Bind("AmenitiesID,RoomID")] RoomAmenities roomAmenities)
+        {
+            if (id != roomAmenities.AmenitiesID)
+            {
+                return NotFound();
+            }
+
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    _context.Update(roomAmenities);
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!RoomAmenitiesExists(roomAmenities.AmenitiesID))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+                return RedirectToAction(nameof(Index));
+            }
+            ViewData["AmenitiesID"] = new SelectList(_context.Amenity, "ID", "Name", roomAmenities.AmenitiesID);
+            ViewData["RoomID"] = new SelectList(_context.Rooms, "ID", "Name", roomAmenities.RoomID);
+            return View(roomAmenities);
+        }
+        
+
+        // GET: RoomAmenities/Delete/5
+        public async Task<IActionResult> Delete(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var roomAmenities = await _context.RoomAmenity
+                .Include(r => r.Amenity)
+                .Include(r => r.Room)
+                .FirstOrDefaultAsync(m => m.AmenitiesID == id);
+            if (roomAmenities == null)
+            {
+                return NotFound();
+            }
+
+            return View(roomAmenities);
+        }
+
+        // POST: RoomAmenities/Delete/5
+        [HttpPost, ActionName("Delete")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> DeleteConfirmed(int id)
+        {
+            var roomAmenities = await _context.RoomAmenity.FindAsync(id);
+            _context.RoomAmenity.Remove(roomAmenities);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
+        }
+
+        private bool RoomAmenitiesExists(int id)
+        {
+            return _context.RoomAmenity.Any(e => e.AmenitiesID == id);
+        }
+    }
+}
