+++ AsyncInn/AsyncInn/Views/Amenities/Index.cshtml
@@ -12,6 +12,12 @@
 <p>
     <a asp-action="Create">Create New</a>
 </p>
+<form asp-controller="Amenities" asp-action="Index">
+    <p>
+        Title: <input type="text" name="SearchString">
+        <input type="submit" value="Filter" />
+    </p>
+</form>
 <p>
     <a asp-controller="Home" asp-action="Index">Home</a>
 </p>
