+++ AsyncInn/AsyncInn/Views/Room/Index.cshtml
@@ -11,6 +11,12 @@
 <p>
     <a asp-action="Create">Create New</a>
 </p>
+<form asp-controller="Room" asp-action="Index">
+    <p>
+        Title: <input type="text" name="SearchString">
+        <input type="submit" value="Filter" />
+    </p>
+</form>
 <p>
     <a asp-controller="Home" asp-action="Index">Home</a>
 </p>
